PRSD-703: Move service tests
